+++ HISTORY.md
@@ -20,6 +20,7 @@
 
 ### 2020-10-07
 * byob: added "settings" button to input slot dialog
+* symbols: added "gearPartial" icon
 
 ### 2020-10-06
 * blocks: disable blurred shadows inside input slots if the hidden "blurred shadows" setting is turned off